Refactored BlogCommentBlockController and BlogItemController to pull logic out into controller services
