+++ src/Foundation/Features/Blog/BlogItem/Index.cshtml
@@ -1,6 +1,6 @@
 ﻿@using EPiServer.Core
 @using EPiServer.Web.Mvc.Html
-@model BlogItemPageModel
+@model BlogItemViewModel
 
 @{
     var contentLoader = EPiServer.ServiceLocation.ServiceLocator.Current.GetInstance<EPiServer.IContentLoader>();
