+++ src/Foundation/Features/Blog/BlogListPage/Preview.cshtml
@@ -3,7 +3,7 @@
 @using EPiServer.Web.Mvc.Html
 @using Foundation.Cms.EditorDescriptors
 
-@model BlogItemPageModel
+@model BlogItemViewModel
 
 @if (string.IsNullOrEmpty(Model.Template) || Model.Template == TemplateSelections.Grid)
 {
